fix: add verification on DeleteBackupEnvironment view

+++ Views/ConsoleViews/ViewStates/DeleteBackupEnvironment.cs
@@ -16,25 +16,45 @@ namespace EasySave.Views.ConsoleViews.ViewStates
             // There is no backup environment available
             if (backupEnvironmentList.Count == 0)
             {
-                Console.Write("There is no backup environment available");
+                Console.WriteLine("There is no backup environment available");
                 Console.WriteLine("");
                 return new MainMenu();
             }
 
-            Console.Write("Select the backup environment you wish to delete:");
-            Console.WriteLine("");
-
-            for (int i = 1; backupEnvironmentList.Count >= i; i++)
+            while (true)
             {
-                Console.WriteLine(i + ". " + backupEnvironmentList[i - 1].Name);
+                // Listing all the backup environments
+                Console.WriteLine("Select the backup environment you wish to delete:");
+                Console.WriteLine("0. Return to menu");
+                for (int i = 1; backupEnvironmentList.Count >= i; i++)
+                {
+                    Console.WriteLine(i + ". " + backupEnvironmentList[i - 1].Name);
+                }
+
+                Console.WriteLine("");
+                string strChoice = Console.ReadLine();
+                int choice;
+
+                if (int.TryParse(strChoice, out choice))
+                {
+                    if (choice == 0)
+                    {
+                        Console.Clear();
+                        return new MainMenu();
+                    }
+
+                    if (choice >= 0 && choice <= backupEnvironmentList.Count)
+                    {
+                        controller.DeleteBackupEnvironment(backupEnvironmentList[choice - 1]);
+
+                        Console.Clear();
+                        Console.WriteLine("The backup environment selected has been deleted");
+                        Console.WriteLine("");
+                        return new MainMenu();
+                    }
+                    Console.Clear();
+                }
             }
-
-            string strChoice = Console.ReadLine();
-            int choice;
-            int.TryParse(strChoice, out choice);
-
-            controller.DeleteBackupEnvironment(backupEnvironmentList[choice - 1]);
-            return new MainMenu();
         }
     }
 }
